requirements.txt
pydantic==2.11.3
requests==2.32.3
crewai-tools==0.42
python-dotenv==1.1.0
boto3==1.38.17
pandas==2.2.3